Extract IDESyncInfo to its own file
